requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
python-dotenv
pickle-mixin
langchain
langchain-community
whitenoise
langchain

